Use same webhook URL for all spaniels
The channel that gets posted to is determined by the config in the tech
docs (`config/tech-docs.yml`) and the `owner_slack` entry in the
documentation source frontmatter so we don't need to specify a separate
Slack webhook URL for each set of documentation we want to scan.

So now, anyone who wants Daniel to look at their docs only needs to add
a `SITE_PAGE_API_URL` entry to the `env.matrix` list.