Only post to the official Slack channels on scheduled runs
Currently, the spaniel runs once a day and posts to the relevant slack
channels defined by `owner_slack` in the tech docs config. It would be
nice to test that adding a site URL to the Travis config here actually
works, so on PRs and merges, Daniel will only post to #bot-testing.