Don't crash if tech docs can't be reached
Currently if Notifier::run is called with a bad URL then it will raise
and the program will crash unless it is caught. Instead we treat this as
a recoverable error; we print to stderr that there was an error, and
don't try to post to Slack, but otherwise don't raise. This means the
Rakefile can continue on to other tech docs.

Co-authored-by: Richard Towers <richard.towers@digital.cabinet-office.gov.uk>